+++ Dockerfile_adapters
@@ -25,7 +25,7 @@ RUN apk --no-cache add \
   && pip3 install --upgrade awscli
 
 # Copy the source code to the working directory
-COPY src ./src
+COPY findr/adapters ./findr/adapters
 
 # Build the TypeScript code
 RUN npm run compile
@@ -35,4 +35,4 @@ RUN npm run compile
 
 # CMD instruction dynamically sets the port based on the adapter type
 # CMD ["sh", "-c", "npm run adapters -- --port $(if [ \"$ADAPTER_TYPE\" = \"http\" ]; then echo 3000; else echo 8883; fi)"]
-CMD ["node", "./dist/adapters.js"]
+CMD ["node", "./dist/adapters/index.js"]
