+++ findr/adapters/README.md
