+++ findr/adapters/processors.ts
@@ -1,6 +1,6 @@
 import fs, { writeFile, existsSync, appendFile } from 'fs'
-import { S3Uploader } from './adapters/s3.js'
-import { DynamoDBUploader } from './adapters/dynamodb.js'
+import { S3Uploader } from './protocols/s3.js'
+import { DynamoDBUploader } from './protocols/dynamodb.js'
 
 /**
  * Represents a handler for transferring data based on source and destination.
