+++ findr/adapters/protocols/dynamodb.ts
