+++ findr/adapters/protocols/s3.ts
